fix: add type casting to int on return

+++ var/www/html/service/data/sanitizer/IntegerSanitizer.php
@@ -5,6 +5,6 @@ namespace App\service\data\sanitizer;
 class IntegerSanitizer implements SanitizerInterface {
     public function sanitize(mixed $value): mixed
     {
-        return filter_var($value, FILTER_SANITIZE_NUMBER_INT);
+        return (int) filter_var($value, FILTER_SANITIZE_NUMBER_INT);
     }
 }
